backend/requirements.txt
---

### 📄 Backend Files

**backend/requirements.txt**
```txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
openai==1.54.0
httpx==0.27.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1